requirements.txt
spotipy>=2.23
python-dotenv>=1.0
scikit-learn>=1.3
numpy>=1.24
pandas>=2.0
matplotlib>=3.7
pytest>=7.4
